Rename StreamReader to BatchStreamReader
