+++ presto-orc/src/main/java/com/facebook/presto/orc/reader/LongBatchStreamReader.java
@@ -19,11 +19,8 @@ import com.facebook.presto.orc.metadata.ColumnEncoding;
 import com.facebook.presto.orc.metadata.ColumnEncoding.ColumnEncodingKind;
 import com.facebook.presto.orc.stream.InputStreamSources;
 import com.facebook.presto.spi.block.Block;
-import com.facebook.presto.spi.type.CharType;
 import com.facebook.presto.spi.type.Type;
-import com.facebook.presto.spi.type.VarcharType;
 import com.google.common.io.Closer;
-import io.airlift.slice.Slice;
 import org.openjdk.jol.info.ClassLayout;
 
 import java.io.IOException;
@@ -31,33 +28,33 @@ import java.io.UncheckedIOException;
 import java.util.List;
 
 import static com.facebook.presto.orc.metadata.ColumnEncoding.ColumnEncodingKind.DICTIONARY;
-import static com.facebook.presto.orc.metadata.ColumnEncoding.ColumnEncodingKind.DICTIONARY_V2;
 import static com.facebook.presto.orc.metadata.ColumnEncoding.ColumnEncodingKind.DIRECT;
 import static com.facebook.presto.orc.metadata.ColumnEncoding.ColumnEncodingKind.DIRECT_V2;
 import static com.facebook.presto.orc.metadata.ColumnEncoding.ColumnEncodingKind.DWRF_DIRECT;
-import static com.facebook.presto.spi.type.Chars.byteCountWithoutTrailingSpace;
-import static com.facebook.presto.spi.type.Chars.isCharType;
-import static com.facebook.presto.spi.type.VarbinaryType.isVarbinaryType;
-import static com.facebook.presto.spi.type.Varchars.byteCount;
-import static com.facebook.presto.spi.type.Varchars.isVarcharType;
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static java.util.Objects.requireNonNull;
 
-public class SliceStreamReader
-        implements StreamReader
+public class LongBatchStreamReader
+        implements BatchStreamReader
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(SliceStreamReader.class).instanceSize();
+    private static final int INSTANCE_SIZE = ClassLayout.parseClass(LongBatchStreamReader.class).instanceSize();
 
     private final StreamDescriptor streamDescriptor;
-    private final SliceDirectStreamReader directReader;
-    private final SliceDictionaryStreamReader dictionaryReader;
-    private StreamReader currentReader;
+    private final LongDirectBatchStreamReader directReader;
+    private final LongDictionaryBatchStreamReader dictionaryReader;
+    private BatchStreamReader currentReader;
 
-    public SliceStreamReader(StreamDescriptor streamDescriptor, AggregatedMemoryContext systemMemoryContext)
+    public LongBatchStreamReader(StreamDescriptor streamDescriptor, AggregatedMemoryContext systemMemoryContext)
     {
         this.streamDescriptor = requireNonNull(streamDescriptor, "stream is null");
-        directReader = new SliceDirectStreamReader(streamDescriptor);
-        dictionaryReader = new SliceDictionaryStreamReader(streamDescriptor, systemMemoryContext.newLocalMemoryContext(SliceStreamReader.class.getSimpleName()));
+        directReader = new LongDirectBatchStreamReader(streamDescriptor, systemMemoryContext.newLocalMemoryContext(LongBatchStreamReader.class.getSimpleName()));
+        dictionaryReader = new LongDictionaryBatchStreamReader(streamDescriptor, systemMemoryContext.newLocalMemoryContext(LongBatchStreamReader.class.getSimpleName()));
+    }
+
+    @Override
+    public void prepareNextRead(int batchSize)
+    {
+        currentReader.prepareNextRead(batchSize);
     }
 
     @Override
@@ -68,26 +65,20 @@ public class SliceStreamReader
     }
 
     @Override
-    public void prepareNextRead(int batchSize)
-    {
-        currentReader.prepareNextRead(batchSize);
-    }
-
-    @Override
     public void startStripe(InputStreamSources dictionaryStreamSources, List<ColumnEncoding> encoding)
             throws IOException
     {
-        ColumnEncodingKind columnEncodingKind = encoding.get(streamDescriptor.getStreamId())
+        ColumnEncodingKind kind = encoding.get(streamDescriptor.getStreamId())
                 .getColumnEncoding(streamDescriptor.getSequence())
                 .getColumnEncodingKind();
-        if (columnEncodingKind == DIRECT || columnEncodingKind == DIRECT_V2 || columnEncodingKind == DWRF_DIRECT) {
+        if (kind == DIRECT || kind == DIRECT_V2 || kind == DWRF_DIRECT) {
             currentReader = directReader;
         }
-        else if (columnEncodingKind == DICTIONARY || columnEncodingKind == DICTIONARY_V2) {
+        else if (kind == DICTIONARY) {
             currentReader = dictionaryReader;
         }
         else {
-            throw new IllegalArgumentException("Unsupported encoding " + columnEncodingKind);
+            throw new IllegalArgumentException("Unsupported encoding " + kind);
         }
 
         currentReader.startStripe(dictionaryStreamSources, encoding);
@@ -106,33 +97,6 @@ public class SliceStreamReader
         return toStringHelper(this)
                 .addValue(streamDescriptor)
                 .toString();
-    }
-
-    public static int getMaxCodePointCount(Type type)
-    {
-        if (isVarcharType(type)) {
-            VarcharType varcharType = (VarcharType) type;
-            return varcharType.isUnbounded() ? -1 : varcharType.getLengthSafe();
-        }
-        if (isCharType(type)) {
-            return ((CharType) type).getLength();
-        }
-        if (isVarbinaryType(type)) {
-            return -1;
-        }
-        throw new IllegalArgumentException("Unsupported encoding " + type.getDisplayName());
-    }
-
-    public static int computeTruncatedLength(Slice slice, int offset, int length, int maxCodePointCount, boolean isCharType)
-    {
-        if (isCharType) {
-            // truncate the characters and then remove the trailing white spaces
-            return byteCountWithoutTrailingSpace(slice, offset, length, maxCodePointCount);
-        }
-        if (maxCodePointCount >= 0 && length > maxCodePointCount) {
-            return byteCount(slice, offset, length, maxCodePointCount);
-        }
-        return length;
     }
 
     @Override
