+++ presto-orc/src/main/java/com/facebook/presto/orc/reader/LongDirectBatchStreamReader.java
@@ -23,6 +23,7 @@ import com.facebook.presto.orc.stream.InputStreamSources;
 import com.facebook.presto.orc.stream.LongInputStream;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
+import com.facebook.presto.spi.block.RunLengthEncodedBlock;
 import com.facebook.presto.spi.type.Type;
 import org.openjdk.jol.info.ClassLayout;
 
@@ -32,17 +33,15 @@ import java.io.IOException;
 import java.util.List;
 
 import static com.facebook.presto.orc.metadata.Stream.StreamKind.DATA;
-import static com.facebook.presto.orc.metadata.Stream.StreamKind.DICTIONARY_DATA;
-import static com.facebook.presto.orc.metadata.Stream.StreamKind.IN_DICTIONARY;
 import static com.facebook.presto.orc.metadata.Stream.StreamKind.PRESENT;
 import static com.facebook.presto.orc.stream.MissingInputStreamSource.missingStreamSource;
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static java.util.Objects.requireNonNull;
 
-public class LongDictionaryStreamReader
-        implements StreamReader
+public class LongDirectBatchStreamReader
+        implements BatchStreamReader
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(LongDictionaryStreamReader.class).instanceSize();
+    private static final int INSTANCE_SIZE = ClassLayout.parseClass(LongDirectBatchStreamReader.class).instanceSize();
 
     private final StreamDescriptor streamDescriptor;
 
@@ -53,24 +52,15 @@ public class LongDictionaryStreamReader
     @Nullable
     private BooleanInputStream presentStream;
 
-    private InputStreamSource<LongInputStream> dictionaryDataStreamSource = missingStreamSource(LongInputStream.class);
-    private int dictionarySize;
-    private long[] dictionary = new long[0];
-
-    private InputStreamSource<BooleanInputStream> inDictionaryStreamSource = missingStreamSource(BooleanInputStream.class);
-    @Nullable
-    private BooleanInputStream inDictionaryStream;
-
-    private InputStreamSource<LongInputStream> dataStreamSource;
+    private InputStreamSource<LongInputStream> dataStreamSource = missingStreamSource(LongInputStream.class);
     @Nullable
     private LongInputStream dataStream;
 
-    private boolean dictionaryOpen;
     private boolean rowGroupOpen;
 
     private LocalMemoryContext systemMemoryContext;
 
-    public LongDictionaryStreamReader(StreamDescriptor streamDescriptor, LocalMemoryContext systemMemoryContext)
+    public LongDirectBatchStreamReader(StreamDescriptor streamDescriptor, LocalMemoryContext systemMemoryContext)
     {
         this.streamDescriptor = requireNonNull(streamDescriptor, "stream is null");
         this.systemMemoryContext = requireNonNull(systemMemoryContext, "systemMemoryContext is null");
@@ -94,14 +84,9 @@ public class LongDictionaryStreamReader
         if (readOffset > 0) {
             if (presentStream != null) {
                 // skip ahead the present bit reader, but count the set bits
-                // and use this as the skip size for the length reader
+                // and use this as the skip size for the data reader
                 readOffset = presentStream.countBitsSet(readOffset);
             }
-
-            if (inDictionaryStream != null) {
-                inDictionaryStream.skip(readOffset);
-            }
-
             if (readOffset > 0) {
                 if (dataStream == null) {
                     throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
@@ -110,83 +95,44 @@ public class LongDictionaryStreamReader
             }
         }
 
-        BlockBuilder builder = type.createBlockBuilder(null, nextBatchSize);
+        if (dataStream == null && presentStream != null) {
+            presentStream.skip(nextBatchSize);
+            Block nullValueBlock = new RunLengthEncodedBlock(
+                    type.createBlockBuilder(null, 1).appendNull().build(),
+                    nextBatchSize);
+            readOffset = 0;
+            nextBatchSize = 0;
+            return nullValueBlock;
+        }
 
+        BlockBuilder builder = type.createBlockBuilder(null, nextBatchSize);
         if (presentStream == null) {
-            // Data doesn't have nulls
             if (dataStream == null) {
                 throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
             }
-            if (inDictionaryStream == null) {
-                for (int i = 0; i < nextBatchSize; i++) {
-                    type.writeLong(builder, dictionary[((int) dataStream.next())]);
-                }
-            }
-            else {
-                for (int i = 0; i < nextBatchSize; i++) {
-                    long id = dataStream.next();
-                    if (inDictionaryStream.nextBit()) {
-                        type.writeLong(builder, dictionary[(int) id]);
-                    }
-                    else {
-                        type.writeLong(builder, id);
-                    }
-                }
-            }
+            dataStream.nextLongVector(type, nextBatchSize, builder);
         }
         else {
-            // Data has nulls
-            if (dataStream == null) {
-                // The only valid case for dataStream is null when data has nulls is that all values are nulls.
-                int nullValues = presentStream.getUnsetBits(nextBatchSize);
-                if (nullValues != nextBatchSize) {
-                    throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
+            for (int i = 0; i < nextBatchSize; i++) {
+                if (presentStream.nextBit()) {
+                    type.writeLong(builder, dataStream.next());
                 }
-                for (int i = 0; i < nextBatchSize; i++) {
+                else {
                     builder.appendNull();
                 }
             }
-            else {
-                for (int i = 0; i < nextBatchSize; i++) {
-                    if (!presentStream.nextBit()) {
-                        builder.appendNull();
-                    }
-                    else {
-                        long id = dataStream.next();
-                        if (inDictionaryStream == null || inDictionaryStream.nextBit()) {
-                            type.writeLong(builder, dictionary[(int) id]);
-                        }
-                        else {
-                            type.writeLong(builder, id);
-                        }
-                    }
-                }
-            }
         }
+
         readOffset = 0;
         nextBatchSize = 0;
+
         return builder.build();
     }
 
     private void openRowGroup()
             throws IOException
     {
-        // read the dictionary
-        if (!dictionaryOpen && dictionarySize > 0) {
-            if (dictionary.length < dictionarySize) {
-                dictionary = new long[dictionarySize];
-            }
-
-            LongInputStream dictionaryStream = dictionaryDataStreamSource.openStream();
-            if (dictionaryStream == null) {
-                throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Dictionary is not empty but data stream is not present");
-            }
-            dictionaryStream.nextLongVector(dictionarySize, dictionary);
-        }
-        dictionaryOpen = true;
-
         presentStream = presentStreamSource.openStream();
-        inDictionaryStream = inDictionaryStreamSource.openStream();
         dataStream = dataStreamSource.openStream();
 
         rowGroupOpen = true;
@@ -195,13 +141,6 @@ public class LongDictionaryStreamReader
     @Override
     public void startStripe(InputStreamSources dictionaryStreamSources, List<ColumnEncoding> encoding)
     {
-        dictionaryDataStreamSource = dictionaryStreamSources.getInputStreamSource(streamDescriptor, DICTIONARY_DATA, LongInputStream.class);
-        dictionarySize = encoding.get(streamDescriptor.getStreamId())
-                .getColumnEncoding(streamDescriptor.getSequence())
-                .getDictionarySize();
-        dictionaryOpen = false;
-
-        inDictionaryStreamSource = missingStreamSource(BooleanInputStream.class);
         presentStreamSource = missingStreamSource(BooleanInputStream.class);
         dataStreamSource = missingStreamSource(LongInputStream.class);
 
@@ -209,7 +148,6 @@ public class LongDictionaryStreamReader
         nextBatchSize = 0;
 
         presentStream = null;
-        inDictionaryStream = null;
         dataStream = null;
 
         rowGroupOpen = false;
@@ -219,14 +157,12 @@ public class LongDictionaryStreamReader
     public void startRowGroup(InputStreamSources dataStreamSources)
     {
         presentStreamSource = dataStreamSources.getInputStreamSource(streamDescriptor, PRESENT, BooleanInputStream.class);
-        inDictionaryStreamSource = dataStreamSources.getInputStreamSource(streamDescriptor, IN_DICTIONARY, BooleanInputStream.class);
         dataStreamSource = dataStreamSources.getInputStreamSource(streamDescriptor, DATA, LongInputStream.class);
 
         readOffset = 0;
         nextBatchSize = 0;
 
         presentStream = null;
-        inDictionaryStream = null;
         dataStream = null;
 
         rowGroupOpen = false;
