+++ presto-orc/src/main/java/com/facebook/presto/orc/reader/MapDirectBatchStreamReader.java
@@ -21,10 +21,11 @@ import com.facebook.presto.orc.stream.BooleanInputStream;
 import com.facebook.presto.orc.stream.InputStreamSource;
 import com.facebook.presto.orc.stream.InputStreamSources;
 import com.facebook.presto.orc.stream.LongInputStream;
-import com.facebook.presto.spi.block.ArrayBlock;
 import com.facebook.presto.spi.block.Block;
+import com.facebook.presto.spi.type.MapType;
 import com.facebook.presto.spi.type.Type;
 import com.google.common.io.Closer;
+import it.unimi.dsi.fastutil.ints.IntArrayList;
 import org.joda.time.DateTimeZone;
 import org.openjdk.jol.info.ClassLayout;
 
@@ -37,20 +38,21 @@ import java.util.Optional;
 
 import static com.facebook.presto.orc.metadata.Stream.StreamKind.LENGTH;
 import static com.facebook.presto.orc.metadata.Stream.StreamKind.PRESENT;
-import static com.facebook.presto.orc.reader.StreamReaders.createStreamReader;
+import static com.facebook.presto.orc.reader.BatchStreamReaders.createStreamReader;
 import static com.facebook.presto.orc.stream.MissingInputStreamSource.missingStreamSource;
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static java.lang.Math.toIntExact;
 import static java.util.Objects.requireNonNull;
 
-public class ListStreamReader
-        implements StreamReader
+public class MapDirectBatchStreamReader
+        implements BatchStreamReader
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(ListStreamReader.class).instanceSize();
+    private static final int INSTANCE_SIZE = ClassLayout.parseClass(MapDirectBatchStreamReader.class).instanceSize();
 
     private final StreamDescriptor streamDescriptor;
 
-    private final StreamReader elementStreamReader;
+    private final BatchStreamReader keyStreamReader;
+    private final BatchStreamReader valueStreamReader;
 
     private int readOffset;
     private int nextBatchSize;
@@ -65,10 +67,11 @@ public class ListStreamReader
 
     private boolean rowGroupOpen;
 
-    public ListStreamReader(StreamDescriptor streamDescriptor, DateTimeZone hiveStorageTimeZone, AggregatedMemoryContext systemMemoryContext)
+    public MapDirectBatchStreamReader(StreamDescriptor streamDescriptor, DateTimeZone hiveStorageTimeZone, AggregatedMemoryContext systemMemoryContext)
     {
         this.streamDescriptor = requireNonNull(streamDescriptor, "stream is null");
-        this.elementStreamReader = createStreamReader(streamDescriptor.getNestedStreams().get(0), hiveStorageTimeZone, systemMemoryContext);
+        this.keyStreamReader = createStreamReader(streamDescriptor.getNestedStreams().get(0), hiveStorageTimeZone, systemMemoryContext);
+        this.valueStreamReader = createStreamReader(streamDescriptor.getNestedStreams().get(1), hiveStorageTimeZone, systemMemoryContext);
     }
 
     @Override
@@ -96,8 +99,9 @@ public class ListStreamReader
                 if (lengthStream == null) {
                     throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
                 }
-                long elementSkipSize = lengthStream.sum(readOffset);
-                elementStreamReader.prepareNextRead(toIntExact(elementSkipSize));
+                long entrySkipSize = lengthStream.sum(readOffset);
+                keyStreamReader.prepareNextRead(toIntExact(entrySkipSize));
+                valueStreamReader.prepareNextRead(toIntExact(entrySkipSize));
             }
         }
 
@@ -105,6 +109,7 @@ public class ListStreamReader
         // and the length values will be converted in-place to an offset vector.
         int[] offsetVector = new int[nextBatchSize + 1];
         boolean[] nullVector = null;
+
         if (presentStream == null) {
             if (lengthStream == null) {
                 throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
@@ -122,32 +127,86 @@ public class ListStreamReader
             }
         }
 
-        // Convert the length values in the offsetVector to offset values in place
+        MapType mapType = (MapType) type;
+        Type keyType = mapType.getKeyType();
+        Type valueType = mapType.getValueType();
+
+        // Calculate the entryCount. Note that the values in the offsetVector are still length values now.
+        int entryCount = 0;
+        for (int i = 0; i < offsetVector.length - 1; i++) {
+            entryCount += offsetVector[i];
+        }
+
+        Block keys;
+        Block values;
+        if (entryCount > 0) {
+            keyStreamReader.prepareNextRead(entryCount);
+            valueStreamReader.prepareNextRead(entryCount);
+            keys = keyStreamReader.readBlock(keyType);
+            values = valueStreamReader.readBlock(valueType);
+        }
+        else {
+            keys = keyType.createBlockBuilder(null, 0).build();
+            values = valueType.createBlockBuilder(null, 1).build();
+        }
+
+        Block[] keyValueBlock = createKeyValueBlock(nextBatchSize, keys, values, offsetVector);
+
+        // Convert the length values in the offsetVector to offset values in-place
         int currentLength = offsetVector[0];
         offsetVector[0] = 0;
         for (int i = 1; i < offsetVector.length; i++) {
-            int nextLength = offsetVector[i];
+            int lastLength = offsetVector[i];
             offsetVector[i] = offsetVector[i - 1] + currentLength;
-            currentLength = nextLength;
+            currentLength = lastLength;
         }
-
-        Type elementType = type.getTypeParameters().get(0);
-        int elementCount = offsetVector[offsetVector.length - 1];
-
-        Block elements;
-        if (elementCount > 0) {
-            elementStreamReader.prepareNextRead(elementCount);
-            elements = elementStreamReader.readBlock(elementType);
-        }
-        else {
-            elements = elementType.createBlockBuilder(null, 0).build();
-        }
-        Block arrayBlock = ArrayBlock.fromElementBlock(nextBatchSize, Optional.ofNullable(nullVector), offsetVector, elements);
 
         readOffset = 0;
         nextBatchSize = 0;
 
-        return arrayBlock;
+        return mapType.createBlockFromKeyValue(Optional.ofNullable(nullVector), offsetVector, keyValueBlock[0], keyValueBlock[1]);
+    }
+
+    private static Block[] createKeyValueBlock(int positionCount, Block keys, Block values, int[] lengths)
+    {
+        if (!hasNull(keys)) {
+            return new Block[] {keys, values};
+        }
+
+        //
+        // Map entries with a null key are skipped in the Hive ORC reader, so skip them here also
+        //
+
+        IntArrayList nonNullPositions = new IntArrayList(keys.getPositionCount());
+
+        int position = 0;
+        for (int mapIndex = 0; mapIndex < positionCount; mapIndex++) {
+            int length = lengths[mapIndex];
+            for (int entryIndex = 0; entryIndex < length; entryIndex++) {
+                if (keys.isNull(position)) {
+                    // key is null, so remove this entry from the map
+                    lengths[mapIndex]--;
+                }
+                else {
+                    nonNullPositions.add(position);
+                }
+                position++;
+            }
+        }
+
+        Block newKeys = keys.copyPositions(nonNullPositions.elements(), 0, nonNullPositions.size());
+        Block newValues = values.copyPositions(nonNullPositions.elements(), 0, nonNullPositions.size());
+        return new Block[] {newKeys, newValues};
+    }
+
+    private static boolean hasNull(Block keys)
+    {
+        for (int position = 0; position < keys.getPositionCount(); position++) {
+            if (keys.isNull(position)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     private void openRowGroup()
@@ -174,7 +233,8 @@ public class ListStreamReader
 
         rowGroupOpen = false;
 
-        elementStreamReader.startStripe(dictionaryStreamSources, encoding);
+        keyStreamReader.startStripe(dictionaryStreamSources, encoding);
+        valueStreamReader.startStripe(dictionaryStreamSources, encoding);
     }
 
     @Override
@@ -192,7 +252,8 @@ public class ListStreamReader
 
         rowGroupOpen = false;
 
-        elementStreamReader.startRowGroup(dataStreamSources);
+        keyStreamReader.startRowGroup(dataStreamSources);
+        valueStreamReader.startRowGroup(dataStreamSources);
     }
 
     @Override
@@ -207,7 +268,8 @@ public class ListStreamReader
     public void close()
     {
         try (Closer closer = Closer.create()) {
-            closer.register(() -> elementStreamReader.close());
+            closer.register(keyStreamReader::close);
+            closer.register(valueStreamReader::close);
         }
         catch (IOException e) {
             throw new UncheckedIOException(e);
@@ -217,6 +279,6 @@ public class ListStreamReader
     @Override
     public long getRetainedSizeInBytes()
     {
-        return INSTANCE_SIZE + elementStreamReader.getRetainedSizeInBytes();
+        return INSTANCE_SIZE + keyStreamReader.getRetainedSizeInBytes() + valueStreamReader.getRetainedSizeInBytes();
     }
 }
